PR code cleanup
